Move DrilldownContext type from reducer to actions file
