+++ app/actions/pageMenu.ts
@@ -5,10 +5,10 @@ import log from 'loglevel';
 import { stringify } from 'query-string';
 import type { SemanticICONS } from 'semantic-ui-react';
 import type { GetInitialTemplateIdResponse } from '../../backend/routes/Templates.types';
-import type { DrilldownContext } from '../reducers/drilldownContextReducer';
 import { NO_PAGES_FOR_TENANT_ERR } from '../utils/ErrorCodes';
 import Internal from '../utils/Internal';
 import { clearContext } from './context';
+import type { DrilldownContext } from './drilldownContext';
 import { popDrilldownContext } from './drilldownContext';
 import type { PageDefinition } from './page';
 import { addLayoutToPage } from './page';
