+++ app/components/Home.tsx
@@ -9,10 +9,10 @@ import type { RouteComponentProps } from 'react-router';
 import type { Dispatch } from 'redux';
 import { storeCurrentPageId, setAppError } from '../actions/app';
 import { clearContext, setValue } from '../actions/context';
+import type { DrilldownContext } from '../actions/drilldownContext';
 import { setDrilldownContext } from '../actions/drilldownContext';
 import { createPagesMap } from '../actions/pageMenu';
 import type { ReduxState } from '../reducers';
-import type { DrilldownContext } from '../reducers/drilldownContextReducer';
 import Consts from '../utils/consts';
 import ContactDetailsModal from './ContactDetailsModal';
 import GettingStartedModal from './GettingStartedModal';
